Conflicts:

	src/utils_cache.c

+++ src/mbmon.c
@@ -240,7 +240,7 @@ static void trim_spaces (char *s)
 {
 	size_t l;
 
-	for (l = strlen (s) - 1; (l > 0) && isspace (s[l]); l--)
+	for (l = strlen (s) - 1; (l > 0) && isspace ((int) s[l]); l--)
 		s[l] = '\0';
 }
